split airportcalc into its own page
