+++ airportcalc.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Airportcalc 2</title>
+    <meta
+      name="description"
+      content="A city and airport size calculator for the Minecart Rapid Transit server"
+    />
+    <meta name="author" content="MRT Mapping Services" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+
+    <link rel="icon" type="image/png" href="./media/ac192.png" />
+    <link rel="manifest" href="manifest.json">
+
+    <link rel="stylesheet" href="out-ac.css" />
+    <script defer src="out-ac.js"></script>
+  </head>
+
+  <body>
+    <div id="map"></div>
+    <div id="bar"></div>
+    <a id="downloader" style="display: none"></a>
+    <input
+      type="file"
+      name="importer"
+      id="importer"
+      accept=".apc,.json"
+      style="display: none"
+    />
+  </body>
+</html>
