+++ index.html
@@ -12,8 +12,8 @@
     <link rel="icon" type="image/png" href="./media/map192.png" />
     <link rel="manifest" href="manifest.json">
 
-    <link rel="stylesheet" href="out.css" />
-    <script defer src="out.js"></script>
+    <link rel="stylesheet" href="out-map.css" />
+    <script defer src="out-map.js"></script>
   </head>
 
   <body>
@@ -29,14 +29,5 @@
     </div>
 
     <div id="map"></div>
-    <div id="bar"></div>
-    <a id="downloader" style="display: none"></a>
-    <input
-      type="file"
-      name="importer"
-      id="importer"
-      accept=".apc,.json"
-      style="display: none"
-    />
   </body>
 </html>
